Repair reference error for goog/base.js.
Repair error reported when loading the page: cannot reference
goog/base.js.cannot reference goog/base.js.

This repair reveals another error. Although the page now loads
correctly, it fails to print the "hello" message to the console.

This error occurs because out generated main.js contains the required
dependency information for the script but **does not** contain actually
invoke the function in the generated script. (It simply includes it in
the page **without** invoking it.)