Write script to create a browser REPL.
Write a Clojure script to create a browser REPL (brepl). Change the core
to create the (one-time) connection to the browser. This change allows
one to evaluate ClojureScript code in the context of the browser.

In addition, this process creates an automatic build process that
watches the source code directory for changes. This allows you to reload
the namespace whenever you make changes and then evaluate newly added
functions in the context of the browser.