Change core module to use multiplication.
Change the code module function `hello-world.core/foo` to multiply its
arguments instead of adding them. By requiring the namespace with the
:reload option, it reloads the namespace allowing us afterward to invoke
the **changed** version of `hello-world.core/foo` in th REPL connected
to our browser.

In addition, the default source map allows the browser and other tools
to report errors against the original ClojureScript instead of the
generated JavaScript.